requirements.txt
django == 4.0.4
python-decouple == 3.6
python-dotenv == 0.20.0
psycopg2-binary == 2.9.3
django-filter == 22.1
djangorestframework == 3.13.1 